Fixes #131: Update test deps.
git-svn-id: https://svn.clinovo.com/clincapture/trunk@86 14881f01-568e-124b-961a-bb94932aa8d1

+++ core/src/test/java/com/clinovo/clincapture/bean/core/UtilsTest.java
@@ -124,13 +124,13 @@ public class UtilsTest {
 
 	@Test
 	public void testGetAge() {
-		assertEquals(10, Utils.getAge(subjectDate, eventDate));
-		assertEquals(10, Utils.getAge(olderSubjectDate, eventDate));
+		assertEquals(10, Utils.getAge(subjectDate, eventDate).intValue());
+		assertEquals(10, Utils.getAge(olderSubjectDate, eventDate).intValue());
 	}
 
 	@Test
 	public void testGetPast() {
-		assertEquals(-1, Utils.getAge(olderSubjectDate, subjectDate));
+		assertEquals(-1, Utils.getAge(olderSubjectDate, subjectDate).intValue());
 	}
 
 	@Test
